third_party/blink/web_tests/http/tests/images/document-policy-oversized-images-responsive-image-expected.txt

CONSOLE ERROR: Document policy violation: oversized-images is not allowed in this document.
